App/requirements.txt
numpy==2.1.2
openai==1.52.2
pandas==2.2.3

psycopg2_binary==2.9.9
python-dotenv==1.0.1
scikit_learn==1.5.2
streamlit==1.39.0
